Ändere Rückgabewert von to_int für PieceType
Ändert die Funktion `to_int` für `PieceType`, um den Rückgabewert von `int` auf `int8_t` zu ändern. Dies könnte den Wertebereich einschränken und Speicherplatz sparen.

Ändere Rückgabewert von to_int für PieceType

Die Funktion `to_int` für den `PieceType` wurde aktualisiert, um den Rückgabewert von `static_cast<int>` auf `static_cast<int8_t>` zu ändern. Diese Änderung könnte den Speicherbedarf reduzieren und die Werte auf einen bestimmten Bereich beschränken.

Ändere Rückgabewert von to_int für PieceType

Die Funktion `to_int` für den `PieceType` wurde aktualisiert, um den Rückgabewert von `static_cast<int>` auf `static_cast<int8_t>` zu ändern. Dies könnte den Speicherbedarf reduzieren und die Werte in einem bestimmten Bereich einschränken.

Ändere Rückgabewert von to_int für PieceType

Ändert die Funktion `to_int` für `PieceType`, um den Rückgabewert von `static_cast<int>` auf `static_cast<int8_t>` zu ändern. Dies könnte den Speicherbedarf reduzieren und die Werte in einem bestimmten Bereich einschränken.

+++ include/utils.h
@@ -66,7 +66,7 @@ inline int to_int(Color color){
 }
 
 inline int to_int(PieceType piece_type){
-    return static_cast<int>(piece_type);
+    return static_cast<int8_t>(piece_type);
 }
 
 inline void remove_castling_right(std::string& rights, char right_to_remove){
